Remove SourceAPI, move source / highlight stuff to compiler module
As we load the compiler stuff in userspace with --scala X.Y.Z, we can
move the source-related stuff along the compiler, and make it live
entirely in userspace, without a bridge to cross the Ammonite /
userspace barrier.
