+++ amm/compiler/src/main/scala/ammonite/repl/Highlighter.scala
