+++ amm/compiler/src/main/scala/ammonite/repl/tools/HighlightJava.scala
