+++ amm/compiler/src/main/scala/ammonite/repl/tools/source.scala
@@ -1,6 +1,5 @@
 package ammonite.repl.tools
 
-import ammonite.repl.api.Location
 import ammonite.runtime.tools.browse.Strings
 import ammonite.util.CodeColors
 import sourcecode.Compat._
@@ -53,7 +52,7 @@ object source{
                   colors: c.Expr[CodeColors]): c.Expr[Unit] = {
     import c.universe._
     val defaultBrowseExpr = c.Expr[Int => Strings](
-      q"_root_.ammonite.repl.api.SourceBridge.value.browseSourceCommand"
+      q"${prefix(c)}.browseSourceCommand"
     )
 
     applyCustomizeCommandMacro(c)(f, defaultBrowseExpr)(pprinter, colors)
@@ -86,7 +85,7 @@ object source{
 
   def prefix(c: Context) = {
     import c.universe._
-    q"ammonite.repl.api.SourceBridge.value"
+    q"ammonite.repl.tools.SourceRuntime"
   }
   /**
     * Attempts to up an expression, into either a LHS + methodcall + rhs. We
