+++ amm/compiler/src/main/scala/ammonite/runtime/tools/Tools.scala
