--- amm/repl/api/src/main/scala/ammonite/repl/tools/desugar.scala
@@ -1,19 +1,0 @@
-package ammonite.repl.tools
-
-import sourcecode.Compat._
-import scala.language.experimental.macros
-
-
-class Desugared(s: String){
-  override def toString() = s
-}
-object desugar{
-  def transformer(c: Context)(expr: c.Expr[Any]): c.Expr[Desugared] = {
-    import c.universe._
-    c.Expr[Desugared](
-      q"ammonite.repl.api.SourceBridge.value.desugarImpl(${c.universe.showCode(expr.tree)})"
-    )
-  }
-
-  def apply(expr: Any): Desugared = macro transformer
-}
