+++ amm/repl/src/main/scala/ammonite/repl/Repl.scala
@@ -137,11 +137,6 @@ class Repl(compilerManager: CompilerLifecycleManager,
       }
     ),
     (
-      "ammonite.repl.api.SourceBridge",
-      "source",
-      new SourceAPIImpl {}
-    ),
-    (
       "ammonite.repl.api.FrontEndBridge",
       "frontEnd",
       new FrontEndAPIImpl {}
