+++ amm/repl/src/test/scala-2.12/ammonite/unit/SourceTests212.scala
@@ -1,8 +1,7 @@
 package ammonite.unit
 
 
-import ammonite.repl.SourceAPIImpl
-import ammonite.repl.api.{Location, SourceBridge}
+import ammonite.repl.tools.Location
 import utest._
 import ammonite.repl.tools.source.load
 import ammonite.util.Util
@@ -10,8 +9,6 @@ import ammonite.util.Util
 
 object SourceTests212 extends TestSuite{
   val tests = Tests{
-    if (SourceBridge.value0 == null)
-      SourceBridge.value0 = new SourceAPIImpl {}
     def check(loaded: Location, expectedFileName: String, expected: String, slop: Int = 10) = {
