+++ amm/repl/src/test/scala/ammonite/unit/SourceTests.scala
@@ -1,8 +1,7 @@
 package ammonite.unit
 
 
-import ammonite.repl.SourceAPIImpl
-import ammonite.repl.api.{Location, SourceBridge}
+import ammonite.repl.tools.Location
 import utest._
 import ammonite.repl.tools.source.load
 import ammonite.util.Util
@@ -10,9 +9,6 @@ import ammonite.util.Util
 import java.io.InputStream
 object SourceTests extends TestSuite{
   val tests = Tests{
-
-    if (SourceBridge.value0 == null)
-      SourceBridge.value0 = new SourceAPIImpl {}
 
     def check(loaded: Location, expectedFileName: String, expected: String, slop: Int = 10) = {
